# openapi/sandicts-api.json
{
  "openapi": "3.1.0",
  "paths": {
...
    "/auth/sign-out": {
      "post": {
        "description": "Revokes the current auth session and clears the refresh cookie.",
        "operationId": "SignOutController_signOutCurrentSession",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Bearer access token for the current session.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Current auth session signed out"
          },
          "401": {
            "description": "Authentication failed. Codes: `invalid_access_token`.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiErrorResponse"
                    }
                  ],
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "enum": [
                        401
                      ]
                    },
                    "code": {
                      "type": "string",
                      "enum": [
                        "invalid_access_token"
                      ]
                    }
                  }
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded. Codes: `rate_limited`.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiErrorResponse"
                    }
                  ],
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "enum": [
                        429
                      ]
                    },
                    "code": {
                      "type": "string",
                      "enum": [
                        "rate_limited"
                      ]
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Unexpected internal failure. Codes: `internal_error`.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiErrorResponse"
                    }
                  ],
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "enum": [
                        500
                      ]
                    },
                    "code": {
                      "type": "string",
                      "enum": [
                        "internal_error"
                      ]
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Sign out",
        "tags": [
          "Auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "Sandicts API",
    "description": "API documentation generated from Nest controllers and Zod schemas.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "ApiValidationIssue": {
        "type": "object",
        "properties": {
          "path": {
            "type": "array",
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "number"
                }
              ]
            }
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "path",
          "message"
        ],
        "id": "ApiValidationIssue"
      },
      "ApiErrorResponse": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "code": {
            "type": "string",
            "enum": [
              "bad_request",
              "validation_error",
              "rate_limited",
              "unauthorized",
              "invalid_google_credential",
              "invalid_magic_link_token",
              "magic_link_expired",
              "magic_link_already_used",
              "magic_link_superseded",
              "email_delivery_unavailable",
              "invalid_access_token",
              "invalid_refresh_token",
              "refresh_token_expired",
              "refresh_token_reused",
              "refresh_token_revoked",
              "auth_session_inactive",
              "forbidden",
              "account_auth_forbidden",
              "external_identity_conflict",
              "resource_not_found",
              "conflict",
              "business_rule_violation",
              "internal_error"
            ]
          },
          "message": {
            "type": "string"
          },
          "path": {
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
          },
          "requestId": {
            "type": "string"
          },
          "issues": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ApiValidationIssue"
            }
          }
        },
        "required": [
          "statusCode",
          "code",
          "message",
          "path",
          "timestamp",
          "requestId"
        ],
        "id": "ApiErrorResponse"
      },
...
    }
  }
}